+++ client.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/rs/zerolog"
 	"go.mau.fi/util/retryafter"
-	"maunium.net/go/maulogger/v2/maulogadapt"
 
 	"maunium.net/go/mautrix/crypto/backup"
 	"maunium.net/go/mautrix/event"
@@ -49,17 +48,6 @@ type VerificationHelper interface {
 	ConfirmSAS(ctx context.Context, txnID id.VerificationTransactionID) error
 }
 
-// Deprecated: switch to zerolog
-type Logger interface {
-	Debugfln(message string, args ...interface{})
-}
-
-// Deprecated: switch to zerolog
-type WarnLogger interface {
-	Logger
-	Warnfln(message string, args ...interface{})
-}
-
 // Client represents a Matrix client.
 type Client struct {
 	HomeserverURL *url.URL     // The base homeserver URL
@@ -75,8 +63,6 @@ type Client struct {
 	Verification  VerificationHelper
 
 	Log zerolog.Logger
-	// Deprecated: switch to the zerolog instance in Log
-	Logger Logger
 
 	RequestHook  func(req *http.Request)
 	ResponseHook func(req *http.Request, resp *http.Response, duration time.Duration)
@@ -2295,7 +2281,7 @@ func NewClient(homeserverURL string, userID id.UserID, accessToken string) (*Cli
 	if err != nil {
 		return nil, err
 	}
-	cli := &Client{
+	return &Client{
 		AccessToken:   accessToken,
 		UserAgent:     DefaultUserAgent,
 		HomeserverURL: hsURL,
@@ -2307,7 +2293,5 @@ func NewClient(homeserverURL string, userID id.UserID, accessToken string) (*Cli
 		// The client will work with this storer: it just won't remember across restarts.
 		// In practice, a database backend should be used.
 		Store: NewMemorySyncStore(),
-	}
-	cli.Logger = maulogadapt.ZeroAsMau(&cli.Log)
-	return cli, nil
+	}, nil
 }